refactor coarsening functions to new module
